refactor and refine docs
